fixes after merge
